+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - [FIX] Text view is now disabled when disabling view.
 - [FIX] When utility button image was set multiple times in a row to an image or
   to nil, it would mess up the view.
+- [IMPROVEMENT] Make use of UI_APPEARANCE_SELECTOR for buttonTintColor,
+  buttonTitle, placeholder, and utilityButtonImage.
 
 ## v1.0.0